Changes to support point of sale

+++ pe-engines-taxes-common/src/main/java/com/threevictors/aws/priceeye/taxes/model/pfcengine/response/PfcResponse.java
@@ -8,5 +8,5 @@ import java.util.List;
 public class PfcResponse{
     public List<Charge> charges;
     public String responseCurrency;
-    public int exchangeRate;
+    public double exchangeRate;
 }
